+++ menu/main.gd
@@ -5,15 +5,24 @@ extends Control
 
 signal menu_changed
 
+func start_game() -> void:
+	var error := get_tree().change_scene("res://game/game.tscn")
+
+	if error != OK:
+		push_error("Couldn't load the game scene.")
+
+func _ready() -> void:
+	# Start the game immediately if `--level=N` is passed on the command line
+	# Setting the starting level is handled in `game/game.gd`
+	if CommandLine.arguments.get("level"):
+		start_game()
+
 func _on_play_pressed() -> void:
 	emit_signal("menu_changed")
 	Music.fade_out()
 
 	yield($"/root/Menu", "transition_finished")
-	var error := get_tree().change_scene("res://game/game.tscn")
-
-	if error != OK:
-		push_error("Couldn't load the game scene.")
+	start_game()
 
 func _on_options_pressed() -> void:
 	emit_signal("menu_changed", $"/root/Menu/Control/Options")