+++ CHANGELOG.md
@@ -9,6 +9,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Screenshots can now be taken by pressing <kbd>F12</kbd>.
+- The game can now be started on a specific level by specifying `--level=N`
+  on the command line, `N` being the number of the level starting from 1.
 
 ## [1.2.0] - 2019-03-01